python/telegram-bot/requirements.txt
axonfi>=0.2.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
